requirements.txt
python-decouple==3.8
mysqlclient==2.2.0
SQLAlchemy==2.0.21